Rearrange code to be more organized
