+++ src/components/GenericInput/types.ts
@@ -1,8 +1,10 @@
-export type SelectOption = {
-  value: string | number;
-  label: string;
-};
+// export type SelectOption = {
+//   value: string | number;
+//   label: string;
+// };
 
+import { SelectOption } from '../common.types';
+export type { SelectOption };
 export type SwitchOption = {
   label: string;
   controlId: string; // for DOM purposes used in <input name,id
