+++ src/components/QueryInput/QInputRange/QInputRange.tsx
@@ -1,4 +1,4 @@
-import { GInputPairSideBySide, Subfield, InputDataType } from '../../GenericInput';
+import { GInputPairSideBySide, Subfield } from '../../GenericInput';
 import { IQInputChange, TermValueChangeMessageOrNull } from '../term.types';
 import { presetOptions } from './presetOptions';
 
@@ -7,21 +7,17 @@ export const untestables = {
   onChangeNoOp,
 };
 
-interface QInputRangeProps extends IQInputChange {
+export interface QInputRangeProps extends IQInputChange {
   rangeOption?: 'inclusive' | 'exclusive';
   expanded?: boolean; //undocumented exposed for test/dev/debug -not expected to be externally
+  // TODO - reorganize Range, create new MinMax (think I deleted when it was called Pair)
+  // TODO - min/max will support {min:null, max:null} where as Range does not
+  // TODO - separate/specialization/generalization
   formatCallbackValues?: (min: any, max: any) => any;
+  // formatCallbackValues?: (min: any, max: any) => TermValue;
   formatDisplayValues?: (min: any, max: any) => string;
-  inputProps?: object;
-  inputDataType?: InputDataType;
-  label?: string;
-
-  // onChange?: (termValue: QueryTermValueOrNull) => void;
   subfields?: { min: Subfield; max: Subfield };
 }
-
-// TODO - add testing (probably borrow heavily from QInputPair)
-// TODO - add InputDataType to testing
 
 type ChangeValue = { [subfieldId: string]: string | number | null };
