+++ src/components/QueryInput/index.ts
@@ -2,4 +2,20 @@ import { QInputSelectSingle, QInputSelectMultiple } from './QInputSelect';
 import { QInputScalar } from './QInputScalar';
 import { QInputRange } from './QInputRange';
 
+import {
+  // QueryTermOperators,
+  QueryTermValueOrNull,
+  SelectOption,
+  // QFieldDataTypeEnum,
+  // QFieldCollection,
+} from './term.types';
+
 export { QInputSelectSingle, QInputSelectMultiple, QInputRange, QInputScalar };
+
+export type { QueryTermValueOrNull, SelectOption };
+
+// export {
+//   QueryTermOperators,
+//   QFieldDataTypeEnum,
+//   QFieldCollection,
+// }
