+++ src/components/common.types.ts
@@ -1,0 +1,4 @@
+export type SelectOption = {
+  value: string | number;
+  label: string;
+};
